# backend/docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ML Evaluation Platform API",
    "description": "API for object detection and segmentation model evaluation platform",
    "version": "1.0.0",
    "contact": {
      "name": "ML Evaluation Platform"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.ml-eval.cloud",
      "description": "Production server"
    }
  ],
  "paths": {
    "/api/v1/images": {
      "post": {
        "summary": "Upload image files",
        "description": "Upload images to the server for processing and evaluation",
        "operationId": "uploadImages",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  },
                  "dataset_split": {
                    "type": "string",
                    "enum": [
                      "train",
                      "validation",
                      "test"
                    ],
                    "default": "train"
                  }
                },
                "required": [
                  "files"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Images uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageUploadResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      },
      "get": {
        "summary": "List images",
        "description": "Retrieve list of uploaded images with filtering",
        "operationId": "listImages",
        "parameters": [
          {
            "name": "dataset_split",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "train",
                "validation",
                "test"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 1000
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of images",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/images/{image_id}": {
      "get": {
        "summary": "Get image details",
        "description": "Retrieve detailed information about a specific image",
        "operationId": "getImage",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Image details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Image"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/v1/evaluation/compare": {
      "post": {
        "summary": "Compare model performance",
        "description": "Compare performance metrics between multiple models",
        "operationId": "compareModels",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelComparisonRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Model comparison results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelComparisonResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Image": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "filename": {
            "type": "string"
          },
          "file_path": {
            "type": "string"
          },
          "file_size": {
            "type": "integer",
            "format": "int64"
          },
          "format": {
            "type": "string"
          },
          "width": {
            "type": "integer"
          },
          "height": {
            "type": "integer"
          },
          "dataset_split": {
            "type": "string",
            "enum": [
              "train",
              "validation",
              "test"
            ]
          },
          "upload_timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "metadata": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "filename",
          "file_path",
          "file_size",
          "format",
          "width",
          "height",
          "dataset_split",
          "upload_timestamp"
        ]
      },
      "ImageUploadResponse": {
        "type": "object",
        "properties": {
          "uploaded_images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Image"
            }
          },
          "total_count": {
            "type": "integer"
          },
          "success_count": {
            "type": "integer"
          },
          "failed_count": {
            "type": "integer"
          }
        },
        "required": [
          "uploaded_images",
          "total_count",
          "success_count",
          "failed_count"
        ]
      },
      "ImageListResponse": {
        "type": "object",
        "properties": {
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Image"
            }
          },
          "total_count": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          }
        },
        "required": [
          "images",
          "total_count",
          "limit",
          "offset"
        ]
      },
...
      "PerformanceMetric": {
        "type": "object",
        "properties": {
          "metric_type": {
            "type": "string"
          },
          "metric_value": {
            "type": "number",
            "format": "float"
          },
          "threshold": {
            "type": "number",
            "format": "float"
          },
          "class_name": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          }
        },
        "required": [
          "metric_type",
          "metric_value"
        ]
      },
      "ModelComparisonRequest": {
        "type": "object",
        "properties": {
          "model_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "dataset_id": {
            "type": "string",
            "format": "uuid"
          },
          "metric_types": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "mAP",
                "mAP@50",
                "mAP@50:95",
                "IoU",
                "precision",
                "recall",
                "F1",
                "execution_time"
              ]
            }
          }
        },
        "required": [
          "model_ids",
          "dataset_id",
          "metric_types"
        ]
      },
      "ModelComparisonResponse": {
        "type": "object",
        "properties": {
          "comparison_results": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "model_id": {
                  "type": "string",
                  "format": "uuid"
                },
                "model_name": {
                  "type": "string"
                },
                "metrics": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PerformanceMetric"
                  }
                }
              }
            }
          },
          "best_performing_model": {
            "type": "string",
            "format": "uuid"
          },
          "evaluation_timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "comparison_results",
          "evaluation_timestamp"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "error",
          "message",
          "timestamp"
        ]
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "PayloadTooLarge": {
        "description": "Request payload too large",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ]
}